SVM: Eliminate filter_executable_program_accounts from pub interface
